adds option to fetch libs only, updates to README

+++ fetch_deps.sh
@@ -2,30 +2,38 @@
 
 set -e
 
-BOARD=$1
+DEFAULT_BOARD="esp32"
+DEFAULT_UPDATE="all"
+BOARD=${1:-DEFAULT_BOARD}
+UPDATE=${2:-DEFAULT_UPDATE}
 
-if [ -d makeEspArduino ]; then rm -rf makeEspArduino; fi
-if [ -d esp8266-arduino ]; then rm -rf esp8266-arduino; fi
-if [ -d arduino-esp32 ]; then rm -rf arduino-esp32; fi
 if [ -d libs ]; then rm -rf libs; fi
-
-# get makeEspArduino
-git clone https://github.com/plerup/makeEspArduino
-cd makeEspArduino
-git checkout tags/4.18.0
-cp makeEspArduino.mk ../Makefile
-cd ../
-
 mkdir -p libs
 
+if [ $UPDATE != "libs" ]; then
+    if [ -d makeEspArduino ]; then rm -rf makeEspArduino; fi
+    if [ -d esp8266-arduino ]; then rm -rf esp8266-arduino; fi
+    if [ -d arduino-esp32 ]; then rm -rf arduino-esp32; fi
+    # get makeEspArduino
+    git clone https://github.com/plerup/makeEspArduino
+    cd makeEspArduino
+    git checkout tags/4.18.0
+    cp makeEspArduino.mk ../Makefile
+    cd ../
+    if [ -d makeEspArduino ]; then rm -rf makeEspArduino; fi
+fi
+
 if [ $BOARD == "esp8266" ]; then
-    # install esp8266-arduino build environment
-    git clone https://github.com/esp8266/Arduino.git esp8266-arduino
-    cd esp8266-arduino/
-    git checkout tags/2.4.0-rc1
-    cd tools/
-    python get.py
-    cd ../../
+    if [ $UPDATE != "libs" ]; then
+        cp config.mk.esp8266 config.mk
+        # install esp8266-arduino build environment
+        git clone https://github.com/esp8266/Arduino.git esp8266-arduino
+        cd esp8266-arduino/
+        git checkout tags/2.4.0-rc1
+        cd tools/
+        python get.py
+        cd ../../
+    fi
     # install ESPAsyncTCP
     cd libs/
     git clone https://github.com/me-no-dev/ESPAsyncTCP
@@ -35,13 +43,16 @@ if [ $BOARD == "esp8266" ]; then
 fi
 
 if [ $BOARD == "esp32" ]; then
-    # install arduino-esp32 build environment
-    git clone https://github.com/espressif/arduino-esp32 arduino-esp32
-    cd arduino-esp32/
-    git checkout tags/1.0.4
-    cd tools/
-    python get.py
-    cd ../../
+    if [ $UPDATE != "libs" ]; then
+        cp config.mk.esp32 config.mk
+        # install arduino-esp32 build environment
+        git clone https://github.com/espressif/arduino-esp32 arduino-esp32
+        cd arduino-esp32/
+        git checkout tags/1.0.4
+        cd tools/
+        python get.py
+        cd ../../
+    fi
     # install ESPAsyncTCP
     cd libs/
     git clone https://github.com/me-no-dev/AsyncTCP
